requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
aio_pika
